refactor: move generate_python_query_body function to a new module and update references
